Do not install headers for private classes
When merging the accessibility plugin into the widgets library, the
headers were just moved. They should have gotten the _p at that time.

Task-number: QTBUG-47569
