+++ src/widgets/accessible/qaccessiblewidgetfactory_p.h
@@ -36,6 +36,17 @@
 #ifndef QACCESSIBLEWIDGETFACTORY_H
 #define QACCESSIBLEWIDGETFACTORY_H
 
+//
+//  W A R N I N G
+//  -------------
+//
+// This file is not part of the Qt API.  It exists purely as an
+// implementation detail.  This header file may change from version to
+// version without notice, or even be removed.
+//
+// We mean it.
+//
+
 QT_BEGIN_NAMESPACE
 
 QAccessibleInterface *qAccessibleFactory(const QString &classname, QObject *object);
