+++ src/widgets/accessible/qaccessiblewidgets_p.h
@@ -34,6 +34,17 @@
 #ifndef QACCESSIBLEWIDGETS_H
 #define QACCESSIBLEWIDGETS_H
 
+//
+//  W A R N I N G
+//  -------------
+//
+// This file is not part of the Qt API.  It exists purely as an
+// implementation detail.  This header file may change from version to
+// version without notice, or even be removed.
+//
+// We mean it.
+//
+
 #include <QtWidgets/qaccessiblewidget.h>
 
 #ifndef QT_NO_ACCESSIBILITY
