+++ src/widgets/accessible/rangecontrols_p.h
@@ -34,6 +34,17 @@
 #ifndef RANGECONTROLS_H
 #define RANGECONTROLS_H
 
+//
+//  W A R N I N G
+//  -------------
+//
+// This file is not part of the Qt API.  It exists purely as an
+// implementation detail.  This header file may change from version to
+// version without notice, or even be removed.
+//
+// We mean it.
+//
+
 #include <QtWidgets/qaccessiblewidget.h>
 
 QT_BEGIN_NAMESPACE
