+++ src/widgets/accessible/simplewidgets_p.h
@@ -34,6 +34,17 @@
 #ifndef SIMPLEWIDGETS_H
 #define SIMPLEWIDGETS_H
 
+//
+//  W A R N I N G
+//  -------------
+//
+// This file is not part of the Qt API.  It exists purely as an
+// implementation detail.  This header file may change from version to
+// version without notice, or even be removed.
+//
+// We mean it.
+//
+
 #include <QtCore/qcoreapplication.h>
 #include <QtWidgets/qaccessiblewidget.h>
